# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Vendor Service Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendor/register:
    post:
      operationId: registerVendor
      tags: [vendors]
      summary: Register as vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorRegisterInput"
      responses:
        "201":
          description: Vendor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /vendor/login:
    post:
      operationId: vendorLogin
      tags: [vendors]
      summary: Vendor login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /vendor/dashboard:
    get:
      operationId: getVendorDashboard
      tags: [vendors]
      summary: Get vendor dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDashboard"
  /vendor/bookings:
    get:
      operationId: getVendorBookings
      tags: [vendors]
      summary: Get vendor bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vendor bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
  /vendor/bookings/{id}:
    patch:
      operationId: updateVendorBooking
      tags: [vendors]
      summary: Update booking status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusInput"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Browse services
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price, rating, createdAt]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Add a service (vendor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getServiceById
      tags: [services]
      summary: Get service details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (vendor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (vendor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required: [field, message]
      required: [success, message]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    VendorRegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        businessName:
          type: string
      required: [name, email, password, businessName]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [success, token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    VendorDashboard:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalServices:
              type: integer
            totalBookings:
              type: integer
            pendingBookings:
              type: integer
            confirmedBookings:
              type: integer
            completedBookings:
              type: integer
            totalEarnings:
              type: number
          required: [totalServices, totalBookings, pendingBookings, confirmedBookings, completedBookings, totalEarnings]
      required: [success, data]
    Service:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        vendorId:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, price, category, vendorId, rating, reviewCount, createdAt]
    ServiceDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        vendorId:
          type: string
        vendor:
          type: object
          properties:
            id:
              type: string
            businessName:
              type: string
          required: [id, businessName]
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, price, category, vendorId, vendor, rating, reviewCount, createdAt]
    ServiceListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [success, data, pagination]
    CreateServiceInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
      required: [title, description, price, category]
    UpdateServiceInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vendorId:
          type: string
        serviceId:
          type: string
        date:
          type: string
        time:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, vendorId, serviceId, date, time, status, paymentStatus, createdAt]
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [success, data, pagination]
...
    UpdateBookingStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, completed, cancelled]
      required: [status]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
      required: [page, limit, total, pages]