# lib/api-zod/src/generated/types/updateBookingStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Vendor Service Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookingStatusInputStatus } from "./updateBookingStatusInputStatus";

export interface UpdateBookingStatusInput {
  status: UpdateBookingStatusInputStatus;
}